feat: null frame support for al codecs
